create router for public routes
